Add 0.6 branch
