+++ src/client.h
@@ -14,10 +14,20 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
 
-#ifndef __LOGIN_H__
-#define __LOGIN_H__
+#ifndef __CLIENT_H__
+#define __CLIENT_H__
 
-void login_calculate(char *, int, const char *, int);
+void client_init();
+void client_stop();
+
+enum connection client_get_conn();
+const char *client_get_raw_addr();
+
+void client_set_nameserver(const char *cp);
+void client_set_topdomain(const char *cp);
+void client_set_password(const char *cp);
+
+int client_handshake(int dns_fd, int raw_mode, int autodetect_frag_size, int fragsize);
+int client_tunnel(int tun_fd, int dns_fd);
 
 #endif
-
